requirements.txt
environs==9.3.2
python-telegram-bot==13.15
vk-api==11.9.9